requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
requests==2.32.3
python-dotenv==1.0.1
thefuzz==0.22.1
python-Levenshtein==0.26.1
psycopg2-binary==2.9.10
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.157.0
APScheduler==3.10.4
